# practices/practice_04/omnixon/omnixon-library/tests/server_openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Interrupted": {
        "properties": {
          "interrupted": {
            "title": "Interrupted",
            "type": "boolean"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "interrupted"
        ],
        "title": "Interrupted",
        "type": "object"
      },
...
      "RecentUser": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "integer"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_active": {
            "format": "date-time",
            "title": "Last Active",
            "type": "string"
          },
          "messages": {
            "title": "Messages",
            "type": "integer"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "agent_id",
          "external_id",
          "timestamp",
          "last_active",
          "messages"
        ],
        "title": "RecentUser",
        "type": "object"
      },
...
      "User": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "integer"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "agent_id",
          "external_id",
          "timestamp"
        ],
        "title": "User",
        "type": "object"
      },
      "UserCreate": {
        "properties": {
          "external_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "External Id",
            "type": "string"
          }
        },
        "required": [
          "external_id"
        ],
        "title": "UserCreate",
        "type": "object"
      },
      "UserUpdate": {
        "properties": {
          "external_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          }
        },
        "title": "UserUpdate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "search_users_api_v1_users_get",
        "parameters": [
          {
            "description": "The start of the external ID (at least 3 characters: there is no list of all users)",
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "description": "The start of the external ID (at least 3 characters: there is no list of all users)",
              "maxLength": 64,
              "minLength": 3,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "description": "Maximum number of results",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "description": "Maximum number of results",
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "title": "Response Search Users Api V1 Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Find users by the start of their external ID",
        "tags": [
          "Main API"
        ]
      },
      "post": {
        "operationId": "create_user_api_v1_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a new user",
        "tags": [
          "Main API"
        ]
      }
    },
    "/api/v1/users/recent": {
      "get": {
        "description": "Latest first, by the latest message that has not expired (`MESSAGE_TTL_DAYS`, 7 by default). Not a list of all users: a user\nwith no kept messages is not in it, but can still be found by `GET /users?query=`.",
        "operationId": "recent_users_api_v1_users_recent_get",
        "parameters": [
          {
            "description": "Maximum number of users",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Maximum number of users",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecentUser"
                  },
                  "title": "Response Recent Users Api V1 Users Recent Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The users of the agent who wrote lately (their messages are still kept)",
        "tags": [
          "Main API"
        ]
      }
    },
    "/api/v1/users/{user_id}": {
      "delete": {
        "operationId": "delete_user_api_v1_users__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a user",
        "tags": [
          "Main API"
        ]
      },
      "get": {
        "operationId": "get_user_api_v1_users__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a user by external ID",
        "tags": [
          "Main API"
        ]
      },
      "patch": {
        "operationId": "update_user_api_v1_users__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update a user",
        "tags": [
          "Main API"
        ]
      }
    },
...
    "/api/v1/users/{user_id}/interrupt": {
      "post": {
        "description": "Ends the stream of this user with the agent of the token: the stream gets `event: interrupted`, what was said so far is\nsaved to the history (marked `interrupted`) and to the memory, and then this returns. A new request of the same user does\nthe same by itself, so this is for stopping without saying anything new.",
        "operationId": "interrupt_user_api_v1_users__user_id__interrupt_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interrupted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop the answer that is being streamed to a user",
        "tags": [
          "Main API"
        ]
      }
    },
...
  }
}